+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+- allow bacon.toml file at workspace level - Fix #141
+
 <a name="v2.9.0"></a>
 ### v2.9.0 - 2023/06/19
 - export format and path can now be configured